pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp",
    "datasets>=4.0.0",
    "esptool",
    "numpy>=2.4.2",
    "pandas",
    "pillow",
    "pip>=26.0.1",
    "platformio>=6.1.19",
    "pyarrow",
    "pyserial",
    "RPi.GPIO",
]

[tool.uv]
package = true
python-preference = "only-system"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
esp = "esp:main"
cmd = "nodes.cmd:main"
launch-nodes = "nodes.launch:main"
record = "nodes.record:main"
replay = "nodes.replay:main"
monitor = "nodes.monitor:main"
